add uv venv and gitignore
I'm including the uv.lock file in the gitignore as there is a requirements.txt file, not 100% sure if this is the correct decision. I'm also including my working_docs in the gitignore because they're really just me playing around with the api but this may not be the correct decision either.